requirements.txt
asgiref==3.6.0
django==4.1.7
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf_nested_routers==0.93.4
gunicorn==20.1.0
psycopg2==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
django-environ==0.9.0
django-filter==22.1
drf-spectacular==0.25.1
django-cors-headers==3.14.0
flake8
isort
whitenoise==6.4.0
